Move new filesystem APIs to a subpackage - there's quite a few of them now, and I'd rather not clutter the main package with more in the future.
