+++ src/main/java/org/quiltmc/loader/api/filesystem/package-info.java
@@ -1,0 +1,14 @@
+/** New subpackage for quilt's filesystem APIs.
+ * <p>
+ * There are 7 older api files which are not part of this package for backwards compatibility:
+ * <ul>
+ * <li>{@link org.quiltmc.loader.api.CachedFileSystem}</li>
+ * <li>{@link org.quiltmc.loader.api.ExtendedFiles}</li>
+* <li>{@link org.quiltmc.loader.api.ExtendedFileSystem}</li>
+ * <li>{@link org.quiltmc.loader.api.FasterFiles}</li>
+ * <li>{@link org.quiltmc.loader.api.FasterFileSystem}</li>
+ * <li>{@link org.quiltmc.loader.api.MountOption}</li>
+ * <li>{@link org.quiltmc.loader.api.QuiltFileSystems}</li>
+ * </ul>
+ */
+package org.quiltmc.loader.api.filesystem;
